requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorclass==2.2.2
Django==4.0.4
django-cors-headers==3.11.0
django-countries==7.3.2
django-rest-framework==0.1.0
djangorestframework==3.13.1
docopt==0.6.2
idna==3.3
mysqlclient==2.1.0
packaging==21.3
Pillow==9.1.0
pip-upgrader==1.4.15
PyJWT==2.3.0
pyparsing==3.0.8
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
terminaltables==3.1.10
typing-extensions==4.2.0
urllib3==1.26.9

